requirements.txt
pytz
Flask
Authlib
python-dotenv
pillow
protobuf